refactor IRtcHandler & VOIPDecorator into /voip
