+++ app/src/androidTest/java/com/github/fribourgsdp/radio/VoiceOverIpEngineDecoratorTest.kt
@@ -2,13 +2,13 @@ package com.github.fribourgsdp.radio
 
 import android.content.Context
 import android.content.Intent
-import androidx.appcompat.app.AppCompatActivity
 import android.widget.ImageButton
 import androidx.test.core.app.ActivityScenario
 import androidx.test.core.app.ApplicationProvider
 import androidx.test.espresso.Espresso
 import androidx.test.espresso.intent.Intents
 import androidx.test.espresso.intent.matcher.IntentMatchers
+import com.github.fribourgsdp.radio.voip.VoiceIpEngineDecorator
 import io.agora.rtc.IRtcEngineEventHandler
 import io.agora.rtc.RtcEngine
 import junit.framework.Assert.assertTrue
